changed streams from int to double in Transformations class

+++ src/main/Transformations.java
@@ -11,24 +11,28 @@ public class Transformations {
 	
 	public static double mean(Records records, int col) {
 		return records.getCol(col).stream()
-				.mapToInt(obj -> (int) Integer.parseInt((String) obj))
+				.mapToDouble(obj -> (double) Double.parseDouble((String) obj))
 				.average()
 				.getAsDouble();
 	}
 	
 	public static double max(Records records, int col){
 		return records.getCol(col).stream()
-				.mapToDouble(obj -> (int) Integer.parseInt((String) obj))
+				.mapToDouble(obj -> (double) Double.parseDouble((String) obj))
 				.max()
 				.getAsDouble();
 	}
 	
 	public static double min(Records records, int col){
 		return records.getCol(col).stream()
-				.mapToDouble(obj -> (int) Integer.parseInt((String) obj))
+				.mapToDouble(obj -> (double) Double.parseDouble((String) obj))
 				.min()
 				.getAsDouble();
 	}
+//	public static int numDistinctElements(Records records, int col) {
+//		return records.getCol(col).stream()
+//				.mapToDouble(obj -> (double) )
+//	}
